requirements_1.txt
feedparser
python-telegram-bot
openai
beautifulsoup4
